requirements.txt
# App

Django==1.9.5
djangorestframework==3.3.2
slothauth==0.3

# Logging

# DevOps

gunicorn==19.4.5
dj-database-url==0.4.0
django-storages==1.1.8
django-boto==0.3.9

# Testing

factory_boy==2.6.1

# Utility

django-extensions==1.6.1
Werkzeug==0.11.4
